changed samlutil.loadElement from string to inputstream to avoid classloader problems

+++ sp/trunk/test/dk/itst/oiosaml/sp/configuration/ConfigurationHandlerTest.java
@@ -47,7 +47,9 @@ public class ConfigurationHandlerTest extends AbstractServiceTests {
 	
 	@After
 	public void tearDown() throws Exception{
-		FileUtils.forceDelete(homeDir);
+		if (homeDir != null) {
+			FileUtils.forceDelete(homeDir);
+		}
 	}
 
 	@Test
